reddit video with audio
The obvious solution would have been to use the dash/hls streams, but they don't have a cors header so I assume they are not meant to be used.

Instead we use an iframe with the reddit video. To prevent autoplay, we don't load the iframe until the user clicks a play button.
This solution has the drawback of not having a functional fullscreen button.